+++ R/sql-backends.R
@@ -1,7 +1,8 @@
-#' @importFrom dplyr sql_select
+#' @importFrom dplyr sql_select sql
 #' @export
+
 sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
-  group_by = NULL, having = NULL, order_by = NULL, limit = NULL,
+  group_by = NULL, having = NULL, order_by = NULL, limit = NULL, distinct = FALSE,
   offset = NULL, fetch = NULL, is_percent = NULL, into = NULL, ...) {
 
   # REFERENCES --------------------------------------------------------------
@@ -19,16 +20,20 @@ sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
 
   assertthat::assert_that(is.character(select), length(select) > 0L)
 
-  if (!is.null(limit)) {
-    limit <- mssql_top(con, limit, is_percent)
+  if (length(limit) > 0L || length(order_by) > 0L) {
+    # If ordering, then TOP should be specified, if it isn't already,
+    # to ensure query works when query is part of a subquery. See #49
+    limit <- mssql_top(con, limit %||% 100, is_percent %||% FALSE)
   }
 
-  out$select <- build_sql("SELECT ", limit, " ",
+  assertthat::assert_that(is.character(select), length(select) > 0L)
+  out$select <- build_sql("SELECT ",
+    if (distinct) sql("DISTINCT "), limit, " ",
     escape(select, collapse = ", ", con = con), con = con)
 
   # INTO --------------------------------------------------------------------
 
-  if (!is.null(into)) {
+  if (length(into) > 0L) {
     out$into <- build_sql("INTO ", into, con = con)
   }
 
@@ -47,7 +52,7 @@ sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
 
   # GROUP BY ----------------------------------------------------------------
 
-  if (!is.null(group_by)) {
+  if (length(group_by) > 0L) {
     assertthat::assert_that(is.character(group_by), length(group_by) > 0L)
     out$group_by <- build_sql("GROUP BY ",
       escape(group_by, collapse = ", ", con = con))
@@ -55,7 +60,7 @@ sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
 
   # HAVING ------------------------------------------------------------------
 
-  if (!is.null(having)) {
+  if (length(having) > 0L) {
     assertthat::assert_that(assertthat::is.string(having))
     out$having <- build_sql("HAVING ",
       escape(having, collapse = ", ", con = con))
@@ -63,29 +68,29 @@ sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
 
   # ORDER BY ----------------------------------------------------------------
 
-  if (!is.null(order_by)) {
+  if (length(order_by) > 0L) {
     assertthat::assert_that(is.character(order_by), length(order_by) > 0L)
     out$order_by <- build_sql("ORDER BY ",
       escape(order_by, collapse = ", ", con = con))
   }
 
   # Offset
-  if (!is.null(offset)) {
+  if (length(offset) > 0L) {
     # OFFSET/FETCH supported for SQL Server 2012 and higher (i.e. 11 and higher)
     # when using ORDER BY
     # http://stackoverflow.com/questions/187998/row-offset-in-sql-server
     # MSSQL versioning: http://support2.microsoft.com/kb/321185
     # OFFSET/FETCH: http://msdn.microsoft.com/en-us/library/ms188385(v=sql.110).aspx
-    assertthat::assert_that(!is.null(order_by),
+    assertthat::assert_that(length(order_by) > 0L,
       dbGetInfo(con)$db.version >= 11, is.integer(offset), offset >= 0)
     out$offset <- build_sql("OFFSET ", offset, con = con)
   }
 
   # Fetch
-  if (!is.null(fetch)) {
+  if (length(fetch) > 0L) {
     # SQL Server 2012 + equivalent of LIMIT is FETCH (used with OFFSET)
     # offset will be non-NULL if it is set and SQL Server dependency is met.
-    assertthat::assert_that(!is.null(offset), dbGetInfo(con)$db.version >= 11,
+    assertthat::assert_that(length(offset) > 0L, dbGetInfo(con)$db.version >= 11,
       is.integer(fetch), fetch >= 0)
     out$fetch <- build_sql("FETCH ", fetch, " ONLY", con = con)
   }
@@ -95,59 +100,6 @@ sql_select.SQLServerConnection <- function (con, select, from, where = NULL,
   escape(unname(compact(out)), collapse = "\n", parens = FALSE, con = con)
 }
 
-#' @importFrom dplyr sql_join sql sql_escape_ident
-#' @importFrom stats setNames
-#' @export
-sql_join.SQLServerConnection <- function(con, x, y, type = "inner",
-  by = NULL, ...) {
-  join <- switch(type,
-    left = sql("LEFT"),
-    inner = sql("INNER"),
-    right = sql("RIGHT"),
-    full = sql("FULL"),
-    stop("Unknown join type:", type, call. = FALSE)
-  )
-
-  by <- common_by(by, x, y)
-
-  # Ensure tables have unique names
-  x_names <- auto_names(x$select)
-  y_names <- auto_names(y$select)
-  uniques <- unique_names(x_names, y_names, by$x[by$x == by$y])
-
-  if (is.null(uniques)) {
-    sel_vars <- c(x_names, y_names)
-  } else {
-    x <- update(x, select = setNames(x$select, uniques$x))
-    y <- update(y, select = setNames(y$select, uniques$y))
-
-    by$x <- unname(uniques$x[by$x])
-    by$y <- unname(uniques$y[by$y])
-
-    sel_vars <- unique(c(uniques$x, uniques$y))
-  }
-
-  xname <- unique_name()
-  yname <- unique_name()
-  on <- sql_vector(paste0(
-    paste0(sql_escape_ident(con, xname), ".",
-      sql_escape_ident(con, by$x)), " = ",
-    paste0(sql_escape_ident(con, yname), ".",
-      sql_escape_ident(con, by$y)),
-    collapse = " AND "), parens = TRUE)
-  cond <- build_sql("ON ", on, con = con)
-
-  from <- build_sql(
-    'SELECT * FROM ',
-    sql_subquery(con, x$query$sql, xname), "\n\n",
-    join, " JOIN \n\n" ,
-    sql_subquery(con, y$query$sql, yname), "\n\n",
-    cond, con = con
-  )
-  attr(from, "vars") <- lapply(sel_vars, as.name)
-
-  from
-}
 
 #' @importFrom dplyr escape
 mssql_top <- function (con, n, is_percent = NULL) {
@@ -165,4 +117,39 @@ mssql_top <- function (con, n, is_percent = NULL) {
     if (!is_mssql_2000) n <- escape(n, parens = TRUE)
     return(build_sql("TOP ", n, " PERCENT"))
   }
+}
+
+## Math (scalar) functions - no change across versions based on eyeballing:
+# MSSQL 2000: https://technet.microsoft.com/en-us/library/aa258862(v=sql.80).aspx
+# MSSQL 2005: https://technet.microsoft.com/en-us/library/ms177516(v=sql.90).aspx
+# MSSQL 2008: https://technet.microsoft.com/en-us/library/ms177516(v=sql.100).aspx
+# MSSQL 2008(r2): https://technet.microsoft.com/en-us/library/ms177516(v=sql.105).aspx
+# MSSQL 2012: https://technet.microsoft.com/en-us/library/ms177516(v=sql.110).aspx
+
+## Aggregate functions
+# MSSQL 2005: https://technet.microsoft.com/en-US/library/ms173454(v=sql.90).aspx
+# MSSQL 2008: https://technet.microsoft.com/en-US/library/ms173454(v=sql.100).aspx
+# MSSQL 2008r2*: https://technet.microsoft.com/en-US/library/ms173454(v=sql.100).aspx
+# MSSQL 2012*: https://technet.microsoft.com/en-US/library/ms173454(v=sql.110).aspx
+# MSSQL 2014: https://technet.microsoft.com/en-US/library/ms173454(v=sql.120).aspx
+#' @importFrom dplyr sql_translate_env sql_variant sql_translator base_scalar
+#' @importFrom dplyr base_agg sql_prefix base_win
+#' @export
+sql_translate_env.SQLServerConnection <- function (con) {
+  sql_variant(
+    scalar = sql_translator(.parent = base_scalar,
+      # http://sqlserverplanet.com/tsql/format-string-to-date
+      as.POSIXct = function(x) build_sql("CAST(", x, " AS DATETIME)"),
+      # DATE data type only available since SQL Server 2008
+      as.Date = function (x) build_sql("CAST(", x, " AS DATE)")
+    ),
+    aggregate = sql_translator(.parent = base_agg,
+      n = function() sql("COUNT(*)"),
+      mean = sql_prefix('AVG'),
+      sd = sql_prefix("STDEV"),
+      sdp = sql_prefix("STDEVP"),
+      varp = sql_prefix("VARP")
+    ),
+    window = base_win
+  )
 }